requirements.txt
beautifulsoup4==4.12.3
llamaapi==0.1.36
plotly==5.22.0
python-dotenv==1.0.1
sec_edgar_downloader==5.0.2
streamlit==1.34.0
